week1/requirements.txt
# Core dependencies
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional but recommended
plotly>=5.15.0
wordcloud>=1.9.0
scikit-learn>=1.3.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
